+++ src/ui/Windows/MainFile.cpp
@@ -1709,7 +1709,7 @@ int DboxMain::DoExportDB(const StringX &sx_Filename, const UINT nID,
     }
   }
 
-  numExported = OIL.size();
+  numExported = (int)OIL.size();
 
   export_core.SetCurFile(sx_Filename);
   export_core.SetReadOnly(false);
@@ -2068,7 +2068,7 @@ void DboxMain::OnExportAttachment()
   if (csMediaType.Left(5) == L"image") {
     // Should be an image file - but may not be supported by CImage - try..
     // Allocate attachment buffer
-    UINT imagesize = att.GetContentSize();
+    UINT imagesize = (UINT)att.GetContentSize();
     HGLOBAL gMemory = GlobalAlloc(GMEM_MOVEABLE, imagesize);
     ASSERT(gMemory);
